Conflicts:
	configure.ac
	testsuite/lda_u/02-ACBN0.test
	testsuite/lda_u/07-noncollinear.01-U5-gs.inp
	testsuite/symmetries/09-symmetrization_gga.test

+++ src/basis_set/atomic_orbital.F90
@@ -58,7 +58,7 @@ contains
     type(geometry_t), target, intent(in)   :: geo
     type(mesh_t),             intent(in)   :: mesh
     integer,                  intent(in)   :: ia, iorb, ispin
-    integer,                  intent(in)   :: truncation
+    integer(8),               intent(in)   :: truncation
     FLOAT,                    intent(in)   :: threshold
 
     type(species_t), pointer :: spec
